+++ src/routes/Navigation.svelte
@@ -25,6 +25,8 @@
   <a href="/tests/spa-values-disappearing">SPA onUpdate</a>
   <a href="/posted">Posted store</a>
   <a href="/tests/flash-onerror">Flash onError</a>
+  <a href="/tests/reset-component">Reset component 1</a>
+  <a href="/tests/reset-component-2">Reset component 2</a>
 </nav>
 
 <style lang="scss">