+++ CHANGELOG.md
@@ -5,10 +5,11 @@ Headlines: Added, Changed, Deprecated, Removed, Fixed, Security
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.5.4] - 2023-08-17
 
 ### Fixed
 
+- Client-side validation wasn't resetted properly, when a component containing a form was destroyed and mounted again.
 - Removed debug statement left from 1.5.3
 
 ## [1.5.3] - 2023-08-16
@@ -23,7 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Forms in components weren't resetted properly when destroyed and created again.
+- Forms in components weren't resetted properly when destroyed and mounted again.
 
 ## [1.5.1] - 2023-08-09